extract record to its own file
