+++ airbyte-integrations/connectors/destination-jdbc/src/main/java/io/airbyte/integrations/destination/jdbc/copy/s3/S3StreamCopierFactory.java
@@ -5,25 +5,22 @@
 package io.airbyte.integrations.destination.jdbc.copy.s3;
 
 import com.amazonaws.services.s3.AmazonS3;
-import com.fasterxml.jackson.databind.JsonNode;
 import io.airbyte.db.jdbc.JdbcDatabase;
 import io.airbyte.integrations.destination.ExtendedNameTransformer;
 import io.airbyte.integrations.destination.jdbc.SqlOperations;
 import io.airbyte.integrations.destination.jdbc.copy.StreamCopier;
 import io.airbyte.integrations.destination.jdbc.copy.StreamCopierFactory;
-import io.airbyte.integrations.destination.jdbc.copy.s3.S3StreamCopierFactory.Config;
-import io.airbyte.integrations.destination.s3.S3DestinationConfig;
 import io.airbyte.protocol.models.AirbyteStream;
 import io.airbyte.protocol.models.ConfiguredAirbyteStream;
 
-public abstract class S3StreamCopierFactory implements StreamCopierFactory<Config> {
+public abstract class S3StreamCopierFactory implements StreamCopierFactory<S3CopyConfig> {
 
   /**
    * Used by the copy consumer.
    */
   @Override
   public StreamCopier create(final String configuredSchema,
-                             final Config config,
+                             final S3CopyConfig config,
                              final String stagingFolder,
                              final ConfiguredAirbyteStream configuredStream,
                              final ExtendedNameTransformer nameTransformer,
@@ -32,7 +29,7 @@ public abstract class S3StreamCopierFactory implements StreamCopierFactory<Confi
     try {
       final AirbyteStream stream = configuredStream.getStream();
       final String schema = StreamCopierFactory.getSchema(stream.getNamespace(), configuredSchema, nameTransformer);
-      final AmazonS3 s3Client = config.s3Config.getS3Client();
+      final AmazonS3 s3Client = config.s3Config().getS3Client();
 
       return create(stagingFolder, schema, s3Client, db, config, nameTransformer, sqlOperations, configuredStream);
     } catch (final Exception e) {
@@ -47,27 +44,10 @@ public abstract class S3StreamCopierFactory implements StreamCopierFactory<Confi
                                          String schema,
                                          AmazonS3 s3Client,
                                          JdbcDatabase db,
-                                         Config config,
+                                         S3CopyConfig config,
                                          ExtendedNameTransformer nameTransformer,
                                          SqlOperations sqlOperations,
                                          ConfiguredAirbyteStream configuredStream)
       throws Exception;
-
-  /**
-   * S3 copy destinations need an S3DestinationConfig to configure the basic upload behavior. We also
-   * want additional flags to configure behavior that only applies to the copy-to-S3 +
-   * load-into-warehouse portion. Currently this is just purgeStagingData, but this may expand.
-   */
-  public record Config(boolean purgeStagingData, S3DestinationConfig s3Config) {
-
-    public static boolean shouldPurgeStagingData(final JsonNode config) {
-      if (config.get("purge_staging_data") == null) {
-        return true;
-      } else {
-        return config.get("purge_staging_data").asBoolean();
-      }
-    }
-
-  }
 
 }
